jobqueue에 _selfPtrqueue 추가 closeServer 접근제한 바꿈
iocnt같은 역할을 하는 selfptrqueue를 추가함

+++ NetworkLibrary/Room.h
@@ -7,7 +7,7 @@ class Room: public JobQueue
 {
 private:
 	friend class RoomManager;
-	int _updateCnt = 0;
+	LONG _updateCnt = 0;
 	HashSet<SessionInfo::ID>_tryEnterSessions;
 	//List<SessionInfo>_tryLeaveSessions;
 	void ProcessEnter();
